+++ src/main/java/com/kiselev/usersubscription/port/UserController.java
@@ -8,7 +8,6 @@ import org.springframework.web.bind.annotation.*;
 @RestController
 @RequiredArgsConstructor
 @RequestMapping("/users")
-
 public class UserController {
     private final UserService userService;
 
@@ -16,11 +15,11 @@ public class UserController {
     public void create(@RequestBody User user) { userService.create(user); }
 
     @GetMapping("/{id}")
-    public User getById(@PathVariable int id) { return userService.getById(id); }
+    public User getById(@PathVariable Long id) { return userService.getById(id); }
 
     @PutMapping("/{id}")
-    public void update(@PathVariable int id, @RequestBody User user) { userService.update(id, user); }
+    public void update(@PathVariable Long id, @RequestBody User user) { userService.update(id, user); }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id) { userService.delete(id); }
+    public void delete(@PathVariable Long id) { userService.delete(id); }
 }
